+++ Filtering/DirectCoupledFilters.h
@@ -17,7 +17,7 @@
 #ifndef DIRECTCOUPLEDFILTERS_H
 #define DIRECTCOUPLEDFILTERS_H
 #include "Filtering/LowpassPrototypeCoeffs.h"
-#include "Filtering/Network.h"
+#include "Schematic/Network.h"
 #include "Schematic/component.h"
 #include "general.h"
 #include <QPen>
@@ -28,11 +28,9 @@ public:
   virtual ~DirectCoupledFilters();
   DirectCoupledFilters(FilterSpecifications);
   void synthesize();
-  SchematicContent getSchematic() { return Schematic; }
 
 private:
   struct FilterSpecifications Specification;
-  SchematicContent Schematic;
 
   std::deque<double> gi; // Lowpass prototype
