+++ Filtering/EndCoupled.h
@@ -18,7 +18,7 @@
 #define ENDCOUPLED_H
 
 #include "Filtering/LowpassPrototypeCoeffs.h"
-#include "Filtering/Network.h"
+#include "Schematic/Network.h"
 #include "Schematic/component.h"
 #include "general.h"
 #include <QPen>
@@ -29,11 +29,9 @@ public:
   virtual ~EndCoupled();
   EndCoupled(FilterSpecifications);
   void synthesize();
-  SchematicContent getSchematic() { return Schematic; }
 
 private:
   struct FilterSpecifications Specification;
-  SchematicContent Schematic;
 };
 
 #endif // ENDCOUPLED_H
