+++ Filtering/QuarterWaveFilters.h
@@ -19,7 +19,7 @@
 #define QUARTERWAVEFILTERS_H
 
 #include "Filtering/LowpassPrototypeCoeffs.h"
-#include "Filtering/Network.h"
+#include "Schematic/Network.h"
 #include "Schematic/component.h"
 #include "general.h"
 #include <QPen>
@@ -30,11 +30,9 @@ public:
   virtual ~QuarterWaveFilters();
   QuarterWaveFilters(FilterSpecifications);
   void synthesize();
-  SchematicContent getSchematic() { return Schematic; }
 
 private:
   struct FilterSpecifications Specification;
-  SchematicContent Schematic;
 };
 
 #endif // QUARTERWAVEFILTERS_H
