+++ Schematic/node.h
@@ -40,40 +40,37 @@
 
 #ifndef NODE_H
 #define NODE_H
+#include "Schematic/Network.h"
 #include "symbol.h"
-#include "Filtering/Network.h"
 
-class Node : public Symbol
-{
+class Node : public Symbol {
 public:
-    Node(GraphWidget *graphWidget);
-    Node(GraphWidget *graphWidget, struct NodeInfo);
-    void addWire(Wire *Wire);
-    QList<Wire *> Wires() const;
+  Node(GraphWidget *graphWidget);
+  Node(GraphWidget *graphWidget, struct NodeInfo);
+  void addWire(Wire *Wire);
+  QList<Wire *> Wires() const;
 
-    enum { Type = UserType + 1 };
-    int type() const { return Type; }
+  enum { Type = UserType + 1 };
+  int type() const { return Type; }
 
-
-    QRectF boundingRect() const;
-    QPainterPath shape() const;
-    void paint(QPainter *painter, const QStyleOptionGraphicsItem *option, QWidget *widget);
-    QPoint getPortLocation(int);
-    QString getID();
+  QRectF boundingRect() const;
+  QPainterPath shape() const;
+  void paint(QPainter *painter, const QStyleOptionGraphicsItem *option,
+             QWidget *widget);
+  QPoint getPortLocation(int);
+  QString getID();
 
 protected:
-    QVariant itemChange(GraphicsItemChange change, const QVariant &value);
+  QVariant itemChange(GraphicsItemChange change, const QVariant &value);
 
-    void mousePressEvent(QGraphicsSceneMouseEvent *event);
-    void mouseReleaseEvent(QGraphicsSceneMouseEvent *event);
+  void mousePressEvent(QGraphicsSceneMouseEvent *event);
+  void mouseReleaseEvent(QGraphicsSceneMouseEvent *event);
 
-    
 private:
-    QList<Wire *> WireList;
-    QPointF newPos;
-    GraphWidget *graph;
-    QString ID;
-
+  QList<Wire *> WireList;
+  QPointF newPos;
+  GraphWidget *graph;
+  QString ID;
 };
 //! [0]
