+++ UI/FilterDesignTool.h
@@ -36,9 +36,9 @@
 #include "Filtering/DirectCoupledFilters.h"
 #include "Filtering/EllipticFilter.h"
 #include "Filtering/EndCoupled.h"
-#include "Filtering/Network.h"
 #include "Filtering/QuarterWaveFilters.h"
 #include "Filtering/SteppedImpedanceFilter.h"
+#include "Schematic/Network.h"
 
 #define LC_LADDER 0
 #define LC_DIRECT_COUPLED 1
